fix(asset copy): Resolve paths correctly. Break apart files & asyncify
